refactor: switch to pnpm and update dependencies
