--- .storybook/theme.js
@@ -1,4 +1,0 @@
-export const themeOptions = {
-  brandTitle: 'React IntersectionObserver',
-  brandUrl: 'https://github.com/thebuilder/react-intersection-observer',
-};
