+++ storybook/.storybook/manager.ts
@@ -1,13 +1,10 @@
 import { addons } from '@storybook/addons';
 import { themes } from '@storybook/theming';
-import { themeOptions } from './theme';
-import { ignoreErrorMessages } from './utils/ignore-errors';
 
 addons.setConfig({
   theme: {
     ...themes.dark,
-    ...themeOptions,
+    brandTitle: 'React IntersectionObserver',
+    brandUrl: 'https://github.com/thebuilder/react-intersection-observer',
   },
 });
-
-ignoreErrorMessages();
