+++ storybook/.storybook/preview.ts
@@ -1,0 +1,8 @@
+import 'intersection-observer';
+import 'tailwindcss/tailwind.css';
+
+export const parameters = {
+  controls: {
+    expanded: true,
+  },
+};
