+++ storybook/package.json
@@ -1,0 +1,55 @@
+{
+  "name": "storybook",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev": "storybook dev -p 9000",
+    "build": "storybook build --output-dir ../example"
+  },
+  "keywords": [],
+  "author": "Daniel Schmidt",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/thebuilder/react-intersection-observer.git"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "framer-motion": "^10.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intersection-observer": "workspace:*",
+    "tailwindcss": "^3.2.7"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.21.0",
+    "@babel/preset-env": "^7.20.2",
+    "@storybook/addon-a11y": "7.0.0-beta.61",
+    "@storybook/addon-actions": "7.0.0-beta.61",
+    "@storybook/addon-docs": "7.0.0-beta.61",
+    "@storybook/addon-essentials": "7.0.0-beta.61",
+    "@storybook/addon-interactions": "7.0.0-beta.61",
+    "@storybook/addon-links": "7.0.0-beta.61",
+    "@storybook/addons": "7.0.0-beta.61",
+    "@storybook/builder-vite": "7.0.0-beta.61",
+    "@storybook/client-logger": "7.0.0-beta.61",
+    "@storybook/preview-api": "7.0.0-beta.61",
+    "@storybook/react": "7.0.0-beta.61",
+    "@storybook/react-vite": "7.0.0-beta.61",
+    "@storybook/theming": "7.0.0-beta.61",
+    "@storybook/types": "7.0.0-beta.61",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@vitejs/plugin-react": "^3.1.0",
+    "autoprefixer": "^10.4.13",
+    "intersection-observer": "^0.12.2",
+    "@mdx-js/react": "^2.3.0",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.4",
+    "prettier-plugin-tailwindcss": "^0.2.4",
+    "storybook": "7.0.0-beta.61",
+    "storybook-dark-mode": "^2.1.1",
+    "typescript": "^4.9.5",
+    "vite": "^4.1.4"
+  }
+}
