+++ storybook/stories/InView.story.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { CSSProperties } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story } from '@storybook/react';
-import { IntersectionOptions, InView } from '../index';
+import { IntersectionOptions, InView } from '../../src';
 import {
   EntryDetails,
   InViewBlock,
