+++ storybook/stories/Intro.mdx
@@ -1,0 +1,4 @@
+import { Markdown } from '@storybook/addon-docs';
+import Readme from '../README.md?raw';
+
+<Markdown>{Readme}</Markdown>
