+++ storybook/stories/Recipes.mdx
@@ -1,0 +1,4 @@
+import { Markdown } from '@storybook/addon-docs';
+import Recipes from '../../docs/Recipes.md?raw';
+
+<Markdown>{Recipes}</Markdown>
