+++ storybook/tailwind.config.js
@@ -1,4 +1,4 @@
 /** @type {import("@types/tailwindcss/tailwind-config").TailwindConfig } */
 module.exports = {
-  content: ['src/**/*.tsx'],
+  content: ['stories/**/*.tsx'],
 };
